+++ packages/razor/resources/views/components/button.razor.php
@@ -1,0 +1,1 @@
+<button class="{{ $classes }}" type="{{ $type }}"@if ($disabled) disabled@endif{!! $attributes !!}><span class="cm-button__label">{{ $label }}</span></button>